Convert filter alphas to uint8_t and use integer math for RPM

+++ gauge_V4/config_calibration.h
@@ -45,20 +45,19 @@ extern uint8_t FILTER_AV3;            // Filter coefficient for sensor C (12/16)
 // ===== HALL EFFECT SPEED SENSOR PARAMETERS =====
 extern uint16_t REVS_PER_KM;        // Revolutions per kilometer (vehicle-specific)
 extern uint8_t TEETH_PER_REV;         // Teeth per revolution (sensor-specific)
-extern float ALPHA_HALL_SPEED;       // EMA filter coefficient (lower value is more filtered)
-extern float HALL_SPEED_MIN;         // Minimum reportable speed (MPH)
+extern uint8_t ALPHA_HALL_SPEED;    // EMA filter coefficient (0-256): 256=no filter, 128=moderate, 64=heavy
+extern uint8_t HALL_SPEED_MIN;      // Minimum reportable speed in km/h*100 (e.g., 50 = 0.5 km/h)
 
 // ===== ENGINE RPM SENSOR PARAMETERS =====
 // Pulses per engine revolution (for 4-stroke engines: cylinders / 2)
 // Examples: 4-cyl=2, 6-cyl=3, 8-cyl=4, 3-cyl=1.5
 extern float PULSES_PER_REVOLUTION;
 
-// EMA filter coefficient (lower value = more filtered)
-// Range: 0.0 to 1.0 (0.7 balances smoothing and responsiveness)
-extern float ALPHA_ENGINE_RPM;
+// EMA filter coefficient (0-256): 256=no filter, 179=moderate (0.7*256), 128=more filtered
+extern uint8_t ALPHA_ENGINE_RPM;
 
 // Minimum reportable RPM (engine idle ~600-800)
-extern float ENGINE_RPM_MIN;
+extern uint8_t ENGINE_RPM_MIN;
 
 // ===== SPEEDOMETER CALIBRATION =====
 extern uint16_t SPEEDO_MAX;    // Maximum speedometer reading: 100 mph * 100 (stored as integer for precision)
